# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: My Service API
  version: 1.0.0
  description: API documentation for the backend service.
servers:
  - url: http://localhost:8080
paths:
...
  /internal/api/product/{productId}/discount:
    post:
      tags:
        - DiscountController
      summary: Apply discount to a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string

        - name: INTERNAL_API_KEY
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscountRequest'
      responses:
        '200':
          description: Operation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
...
  /api/v1/product/{productId}/comments:
    get:
      summary: List comments
      tags:
        - Comment Controller
      description: Returns paginated list of comments for a product
      parameters:
        - name: productId
          in: path
          description: Product id
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Page size
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentListResponse'
  /api/v1/product/{productId}/comment:
    post:
      summary: Create comment
      tags:
        - Comment Controller
      description: Create a new comment for a product
      parameters:
        - name: productId
          in: path
          description: Product id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Comment payload
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
...
components:
  schemas:
...
    GenericResponse:
      type: object
      properties:
        httpStatus:
          type: integer
        message:
          type: string
...
    DiscountRequest:
      type: object
      description: Discount request
      properties:
        productId:
          type: string
          description: Product ID (required)
        description:
          type: string
          maxLength: 255
          nullable: true
        percentage:
          type: number
          format: double
          minimum: 0.00
          maximum: 100.00
          description: Discount percentage (required)
        startDate:
          type: string
          format: date-time
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
    CommentRequest:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    CommentListResponse:
      type: object
      properties:
        commentList:
          type: array
          items:
            $ref: '#/components/schemas/CommentResponse'
        page:
          type: integer
        size:
          type: integer
        totalPage:
          type: integer
      required:
        - commentList
        - page
        - size
        - totalPage
    CommentResponse:
      type: object
      properties:
        userId:
          type: string
        userName:
          type: string
        profileImage:
          type: string
          nullable: true
        content:
          type: string
      required:
        - userId
        - userName
        - content